Extract /auth routes into flux/api/auth_routes.py (AuthRoutesMixin)
Establishes the route-extraction pattern: route groups become mixin classes
whose _register_*_routes(self, api, *, auth_config, auth_service,
principal_registry, limiter) method defines the handler closures verbatim, so
they keep self and the shared deps. server.py calls them from _create_api.
Security suite green (369 passed).
