+++ flux/server.py
@@ -45,6 +45,7 @@ from flux.domain.schedule import schedule_factory
 from flux.security.auth_service import AuthService
 from flux.security.dependencies import init_auth_service, get_identity, require_permission
 from flux.security.identity import ANONYMOUS, FluxIdentity
+from flux.api.auth_routes import AuthRoutesMixin
 from datetime import datetime, timezone
 
 # Re-exported for backward compatibility (these were defined here before the
@@ -89,7 +90,7 @@ from flux.api.schemas import (  # noqa: F401
 logger = get_logger(__name__)
 
 
-class Server:
+class Server(AuthRoutesMixin):
     """
     Server for managing workflows and tasks with integrated scheduler.
     """
@@ -5082,69 +5083,13 @@ class Server:
             except Exception as e:
                 raise HTTPException(status_code=500, detail=str(e))
 
-        # ===========================================
-        # Auth: Permissions, Test Token
-        # ===========================================
-
-        @api.get("/auth/permissions")
-        async def auth_permissions(
-            workflow: str | None = None,
-            identity: FluxIdentity = Depends(get_identity),
-        ):
-            try:
-                catalog = WorkflowCatalog.create()
-                if workflow:
-                    from flux.catalogs import resolve_workflow_ref as _resolve_perm
-
-                    _perm_ns, _perm_name = _resolve_perm(workflow)
-                    wf = catalog.get(_perm_ns, _perm_name)
-                    meta = wf.metadata or {} if hasattr(wf, "metadata") else {}
-                    perms = [f"workflow:{wf.namespace}:{wf.name}:read"]
-                    perms.extend(
-                        auth_service._collect_required_permissions(
-                            namespace=wf.namespace,
-                            workflow_name=wf.name,
-                            workflow_metadata=meta,
-                        ),
-                    )
-                    return perms
-                result = {}
-                for wf in catalog.all():
-                    meta = wf.metadata or {} if hasattr(wf, "metadata") else {}
-                    perms = [f"workflow:{wf.namespace}:{wf.name}:read"]
-                    perms.extend(
-                        auth_service._collect_required_permissions(
-                            namespace=wf.namespace,
-                            workflow_name=wf.name,
-                            workflow_metadata=meta,
-                        ),
-                    )
-                    result[f"{wf.namespace}/{wf.name}"] = perms
-                return result
-            except Exception as e:
-                raise HTTPException(status_code=500, detail=str(e))
-
-        @api.post("/auth/test-token")
-        @limiter.limit("10/minute")
-        async def auth_test_token(
-            request: Request,
-            body: dict,
-            identity: FluxIdentity = Depends(require_permission("admin:*")),
-        ):
-            try:
-                token = body.get("token")
-                if not token:
-                    return {"valid": False, "error": "Missing 'token' in request body"}
-                tested_identity = await auth_service.authenticate(token)
-                permissions = await auth_service.resolve_permissions(tested_identity)
-                return {
-                    "valid": True,
-                    "subject": tested_identity.subject,
-                    "roles": sorted(tested_identity.roles),
-                    "permissions": sorted(permissions),
-                }
-            except Exception:
-                return {"valid": False, "error": "Invalid or expired token"}
+        self._register_auth_routes(
+            api,
+            auth_config=auth_config,
+            auth_service=auth_service,
+            principal_registry=principal_registry,
+            limiter=limiter,
+        )
 
         # ===========================================
         # Execution Authorization (task callbacks)
